Create Convert int to float
A new upgrade of your favorite coding language has released. It is the 4th version. However, the industry likes to view it in decimal points. Thus 4 should be viewed as 4.0

Store the value 4 in an an integer and then convert it to a float value. Print the float value.

Example Output:

4.0