Create Slice it
You are decoding the agent name of a rival nation. Fortunately, the code is simple to crack. The real name can be obtained by extracting the letters from the 3rd position to the 6th position (both inclusive).

Given an agent_name, extract and print the real name.

Complete the given program