Create Input shoe and size
Write a program the takes as input the shoe and size of the user. Print it in the given format

Example Input

Boots
13
Output

User wears Boots - size is 13
Hint: Are you getting an extra line in your output? Try using the strip() method - it strips the input string from any extra characters.

Try the following:
shoe = input().strip()