Create Input echoing words
Given a word and a number, you must print it's echo. The echo is the repetition of the word for the given number of repetitions.

Example Input

Hello
5
Output

HelloHelloHelloHelloHello


The number will be given to you as an input to the program. Two lines in each input. You have to read these lines with the input() command first, before echo-ing the answer!