Create specialPrint some numbers
A function specialPrint has been defined. It takes as parameter an integer and prints each of the digits in word format.

Example: for input specialPrint(62), it prints: six two

Write a program that uses specialPrint to print the values for 37, 93, 14 and 568.

Note: The method has already been defined. Your job is to just call the method with the given values