Create specialSum of numbers
A function specialSum has been defined. It takes as input three integers and returns the sum of their squares.

Example: for input specialSum(3,4,5), it returns: 50

Write a program that uses specialSum to print the values for:

1,1,1
2,15,79
4,5,6
24,25,23
-1, -3, 4
Note: The function has already been defined. You just need to call it with the given values.