Create funnyPrint in Camel Case
A function funnyPrint has been defined. It takes as input a String and prints the string in Camel Case.

Example: for input funnyPrint("hello mate"), it prints: HelloMate

Write a program that uses funnyPrint to print the values for:

insert delete
there are two ways to LIVE
Programmer 4 lIFE
Note: the function has already been defined. You just need to call it with the given values.