Create change the name, please
A function nameChanger has been defined. It takes as input 2 Strings. The first is the original string, and the second is the substitute string. It returns a string with all occurrences of the substitute string replaced with a % in the original string.

Example: for input nameChanger("hello", "e"), it returns: h%llo

Write a program that uses nameChanger to print the values for:

"vowel", "l"
"superhero", "r"
"Java", "a"
Note: The function has already been defined. You just need to call it with the given values.