Create Powerful 3
A method largestPower has been defined. It takes as input an integer and returns the largest power of 3 less than or equal to the integer.

Example: for input largestPower(28), it returns: 27

Write a program that uses largestPower to print the values for:

81
181
1881
Note: The function has already been defined. You just need to call it with the given values.