Create Return the nth character
Complete the given method called solve that takes as parameters an integer n and a String word.

Your method should return the character which is at the nth position of word.

It returns the value of the character at the nth position. Eg, solve(3,"Earth") returns r.

Only complete the given method. The rest of the program has already been written