Create Reverse the digits
Complete the method called solve that takes as input an integer and returns a reversed integer.

This integer will always be a 2 digit number. You have to return another integer such that the 2 digits are reversed.

Example: 79 returns 97
Example: 81 returns 18

Only complete the given method. The rest of the program has already been written