Create What's the distance?
Complete the method called solve that takes the Cartesian coordinates of two points and returns the distance between the 2 points.

Here (x1,y1) are the coordinates of the first point and (x2,y2) are the coordinates of the second point.

Only complete the given method. The rest of the program has already been written

Note: x1, y1, x2, y2 are integers. The distance to be returned is float type.

Eg, Distance between (1,0) and (0,0) is 1.0
Distance between (0,3) and (4,0) is 5.0